+++ skcriteria/preprocessing/distance.py
@@ -18,7 +18,8 @@
 
 import numpy as np
 
-from ..core import Objective, SKCTransformerABC
+from ._preprocessing_base import SKCTransformerABC
+from ..core import Objective
 from ..utils import doc_inherit
 
 # =============================================================================