+++ skcriteria/preprocessing/increment.py
@@ -24,7 +24,7 @@ to add value to zero on an array along an arbitrary axis.
 
 import numpy as np
 
-from ..core import SKCMatrixAndWeightTransformerABC
+from ._preprocessing_base import SKCMatrixAndWeightTransformerABC
 from ..utils import doc_inherit
 
 # =============================================================================