+++ skcriteria/preprocessing/weighters.py
@@ -29,9 +29,14 @@ import numpy as np
 import scipy.stats
 
 
+from ._preprocessing_base import SKCTransformerABC
 from .distance import cenit_distance
-from ..core import Objective, SKCTransformerABC
+from ..core import Objective
 from ..utils import doc_inherit
+
+# =============================================================================
+# BASE CLASS
+# =============================================================================
 
 
 class SKCWeighterABC(SKCTransformerABC):